requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-wtf==1.2.1
werkzeug==3.0.3
apscheduler==3.10.4
requests==2.32.3
anthropic==0.28.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.2.2
gunicorn==22.0.0
